Moving contexts to top level
