+++ src/motion/features/use-features.tsx
@@ -1,6 +1,6 @@
 import * as React from "react"
 import { useContext } from "react"
-import { MotionConfigContext } from "../context/MotionConfigContext"
+import { MotionConfigContext } from "../../context/MotionConfigContext"
 import { VisualElement } from "../../render/types"
 import { MotionProps } from ".."
 import { MotionFeature } from "./types"
